shop_api/requirements.txt
Django>=4.2
djangorestframework
drf-yasg
django-dotenv
django-jazzmin
psycopg2-binary
